Fixes issue where you can't scroll vertically in a horizontal list. Also refactors horizontal list code to be cleaner.
