+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@protonapp/material-components",
-  "version": "0.8.18",
+  "version": "0.9.0",
   "description": "Material UI Components for Proton",
   "main": "index.js",
   "scripts": {
