+++ src/ChipList/ImageScrollView.js
@@ -1,37 +1,13 @@
 import React, { Component } from 'react'
-import {
-  Text,
-  View,
-  StyleSheet,
-  Image,
-  ScrollView,
-  TouchableWithoutFeedback,
-} from 'react-native'
-import ImageItem from './ImageItem.js'
+import { ScrollView } from 'react-native'
 
 class ImageScrollView extends Component {
   render() {
-    const { imageList, style, rightIcon } = this.props
+    const { children } = this.props
 
     return (
       <ScrollView horizontal={true} showsHorizontalScrollIndicator={false}>
-        <View style={{ flexDirection: 'row', justifyContent: 'flex-start' }}>
-          {imageList &&
-            imageList.map(
-              ({ clickActions, text, image, rightIcon, chipStyles }, index) => (
-                <View style={style.container} key={index}>
-                  <ImageItem
-                    image={image.image}
-                    style={style}
-                    title={text.title}
-                    imageProps={image}
-                    clickActions={clickActions}
-                    rightIcon={rightIcon}
-                  />
-                </View>
-              )
-            )}
-        </View>
+        {children}
       </ScrollView>
     )
   }
