+++ src/HorizontalImageList/ImageScrollView.pwa.js
@@ -1,0 +1,28 @@
+import React, { Component } from 'react'
+import {
+  Text,
+  View,
+  StyleSheet,
+  Image,
+  ScrollView,
+  ImageBackground,
+  TouchableWithoutFeedback,
+} from 'react-native'
+import ImageItem from './ImageItem.js'
+import BottomBar from './BottomBar.js'
+import placeholder from './holdplace.png'
+
+class ImageScrollView extends Component {
+  render() {
+    const { children } = this.props
+
+    return (
+      <ScrollView showsHorizontalScrollIndicator={false}>
+        <View style={{ flexDirection: 'row', overflow: 'auto' }}>
+          {children}
+        </View>
+      </ScrollView>
+    )
+  }
+}
+export default ImageScrollView
